Persisting game state to user

+++ backend/migrations/sqls/20200419000337-set-up-game-saves-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE game_saves(
   id serial PRIMARY KEY,
-  resources BIGINT DEFAULT 0,
+  resources INT DEFAULT 0,
   game_state JSONB,
   created_on TIMESTAMP NOT NULL,
   last_interaction TIMESTAMP,
